Change CNAME records to DNS records
The DNS instructions in RFC1034 (section 3.6.2) state that, if a CNAME record is present at a node, no other data should be present. This ensures the data for a canonical name and its aliases cannot be different. Because an apex domain requires NS records and usually other records like MX to make it work, setting a CNAME at the apex would break the "no other data should be present" rule.
 
As a CNAME record is not allowed to coexist with any other data, having the apex domain as a CNAME can cause issues, for example with MX (mail).
 
The RFC standard is to use type A for apex (ex: example.com) and CNAME for subdomains (ex: www.example.com)