Clarify usage of DNS types
The DNS instructions in RFC1034 state that, if a CNAME record is present at a node, no other data should be present. This ensures the data for a canonical name and its aliases cannot be different. Because an apex domain requires NS records and usually other records like MX to make it work, setting a CNAME at the apex would break the "no other data should be present" rule.

The RFC standard is to use type A for apex (ex: example.com) and CNAME for subdomains (ex: www.example.com)