Suggest documentation changes for github integration
When you use `magento-cloud integration:add --type=github --project=XXXXXX --base-url=https://github.com --token=XXXXXXXXX --repository={account}/{repository-name} ` you might end up with an error like "The integration is invalid.
The following error was found: repository: Unable to connect to GitHub: repository not found". Even though the repository exists and token has all the proper permissions, documentation is misleading because it does not PROPERLY and FULLY mentioning that "base-url" is needed only when repository is on a private server. This needs to be as clear as possible.